feat(lap-711): redirect when create link success and speaking promotion
