+++ .github/workflows/deploy.yml
@@ -1,6 +1,13 @@
 name: Deploy React App to DigitalOcean
 
 on:
+  workflow_dispatch:
+    inputs:
+      branch:
+        description: 'Branch to deploy'
+        required: true
+        default: 'main'
+
   push:
     branches:
       - main
