+++ src/components/pages/result/speaking-result.tsx
@@ -1,3 +1,4 @@
+import { SubscriptionPromotion } from "@components/molecules/subscription-promotion.tsx";
 import { useTranslation } from "react-i18next";
 
 import icons from "@/assets/icons";
@@ -9,7 +10,7 @@ import { useGetSimulatedTestDetail } from "@/hooks/react-query/use-simulated-tes
 import { MAPPED_SPEAKING_CRITERIA_TITLES } from "@/lib/consts";
 import { EnumSimulatedTestSessionStatus, EnumSkill, EnumSpeakingCriteria } from "@/lib/enums";
 import { SpeakingSession } from "@/lib/types/simulated-test-session.type";
-import { formatTime } from "@/lib/utils/index";
+import { cn, formatTime } from "@/lib/utils";
 
 type SpeakingResultProps = {
   session: SpeakingSession;
@@ -18,7 +19,8 @@ type SpeakingResultProps = {
 function SpeakingResult({ session }: SpeakingResultProps) {
   const { data: stData } = useGetSimulatedTestDetail(session.skillTest.simulatedIeltsTest.id);
   const { t } = useTranslation();
-  const questionTypes =
+  // Get part types: Line graph, Pie chart, etc.
+  const partDetails =
     stData?.skillTests
       .find((skill) => skill.skill === EnumSkill.speaking)
       ?.partsDetail.map((item) => item.questionTypes) ?? [];
@@ -32,14 +34,17 @@ function SpeakingResult({ session }: SpeakingResultProps) {
       ) : (
         <div className="rounded-xl bg-white p-4 xl:p-8">{t("crashMessage")}</div>
       )}
-      <SpeakingSubmission
-        userSubmissions={session.responses}
-        skillTestId={session.skillTest.id}
-        questionTypes={questionTypes}
-        evaluationResults={session.results}
-        parts={session.parts}
-        resource={session.resource}
-      />
+      <div className="grid grid-cols-1 gap-2 md:grid-cols-6 md:gap-8">
+        <div className={cn("flex gap-4 pb-0", !session.results ? "col-span-4" : "col-span-full")}>
+          <SpeakingSubmission
+            userSubmissions={session.responses}
+            skillTestId={session.skillTest.id}
+            partDetails={partDetails}
+            evaluationResults={session.results}
+          />
+        </div>
+        <SubscriptionPromotion results={session.results} />
+      </div>
     </div>
   );
 }
